Revert "Spots final"
