--- babel.config.js
@@ -1,12 +1,0 @@
-const presets = [
-    [
-        "@babel/preset-env",
-        {
-            targets: "defaults, IE 11, not dead",
-            useBuiltIns: "entry",
-            corejs: "^3",
-        },
-    ],
-];
-
-module.exports = { presets };
