+++ blocks/modal.css
@@ -1,0 +1,192 @@
+.modal {
+    font-family: "Poppins", Arial, Helvetica, sans-serif;
+    line-height: 1;
+    font-size: 16px;
+    font-weight: 500;
+    background-color: rgba(0, 0, 0, 0.5);
+    padding: 0 12px;
+    position: fixed;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    visibility: hidden;
+    opacity: 0;
+    z-index: -1;
+    transition: all 0.25s ease;
+}
+
+.modal_opened {
+    visibility: visible;
+    z-index: 3;
+    opacity: 1;
+    transition: all 0.25s ease;
+}
+
+.modal__content {
+    position: relative;
+    max-width: 730px;
+    width: 100%;
+    box-sizing: border-box;
+    padding: 32px 50px;
+    background: #fff;
+    border-radius: 8px;
+}
+
+.modal__content_type_preview {
+    padding: 0;
+    background: transparent;
+    display: flex;
+    flex-direction: column;
+}
+
+.modal__img {
+    width: 100%;
+    max-width: 80vw;
+    max-height: 80vh;
+    object-fit: contain;
+    border-radius: 8px;
+}
+
+.modal__heading {
+    margin: 0 0 32px;
+    padding: 0;
+    max-width: 630px;
+    font-size: 32px;
+    font-weight: 500;
+}
+
+.modal__close-btn {
+    position: absolute;
+    padding: 0;
+    width: 40px;
+    height: 40px;
+    background: transparent url(../images/close-button.svg) center no-repeat;
+    background-size: 35px 35px;
+    top: 32px;
+    right: 50px;
+    opacity: 0.4;
+    border: none;
+}
+
+.modal__close-btn_type_preview {
+    background-image: url(../images/close-button-white.svg);
+    opacity: 1;
+    top: 0;
+    right: -52.5px;
+}
+
+.modal__close-btn:hover {
+    opacity: 1;
+    cursor: pointer;
+}
+
+.modal__caption {
+    position: relative;
+    color: #fff;
+    background-color: #212121;
+    line-height: 1.2;
+    font-size: 20px;
+    font-weight: 500;
+    padding: 4px 8px;
+    margin: 8px 0 0;
+    border-radius: 8px;
+    align-self: start;
+}
+
+.modal__form {
+    display: flex;
+    flex-direction: column;
+    gap: 32px;
+}
+
+.modal__label {
+    color: rgba(21, 21, 21, 0.7);
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+    line-height: 1.1;
+}
+
+.modal__input {
+    border: 1px solid rgba(21, 21, 21, 0.4);
+    padding: 16px 12px;
+    color: rgba(21, 21, 21, 0.3);
+    border-radius: 8px;
+    font-size: 20px;
+}
+
+.modal__input:hover,
+.modal__input:focus {
+    border: 1px solid rgba(21, 21, 21, 1);
+}
+
+.modal__input:not(:placeholder-shown) {
+    color: rgba(21, 21, 21, 1);
+}
+
+.modal__input_invalid,
+.modal__input_invalid:hover,
+.modal__input_invalid:focus {
+    border: 1px solid rgba(255, 0, 0, 0.6);
+}
+
+.modal__error {
+    color: rgba(255, 0, 0, 0.7);
+    line-height: 1.1;
+}
+
+.modal__save-btn {
+    padding: 18px;
+    border-radius: 8px;
+    background-color: #000;
+    width: 156px;
+    border: none;
+    color: #fcf5e5;
+}
+
+.modal__save-btn:hover {
+    background-color: rgba(0, 0, 0, 0.8);
+    cursor: pointer;
+}
+
+.modal__save-btn_disabled {
+    background-color: rgba(0, 0, 0, 0.5);
+    pointer-events: none;
+}
+
+@media screen and (max-width: 627px) {
+    .modal {
+        padding: 0 10px;
+    }
+
+    .modal__content {
+        padding: 12px;
+    }
+
+    .modal__image {
+        border-radius: 4px;
+    }
+
+    .modal__title {
+        font-size: 20px;
+    }
+
+    .modal__close-btn {
+        top: 12px;
+        right: 12px;
+    }
+
+    .modal__close-btn_type_preview {
+        order: -1;
+        right: 13px;
+        top: -36px;
+    }
+
+    .modal__save-btn {
+        width: 100%;
+    }
+}
