--- package.json
@@ -1,45 +1,0 @@
-{
-  "name": "se_project_spots",
-  "version": "1.0.0",
-  "description": "Photo sharing website",
-  "main": "index.js",
-  "scripts": {
-    "build": "webpack --mode production",
-    "dev": "webpack serve --mode development",
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d dist"
-  },
-  "devDependencies": {
-    "@babel/preset-env": "7.27.2",
-    "autoprefixer": "10.4.21",
-    "babel-loader": "10.0.0",
-    "clean-webpack-plugin": "4.0.0",
-    "css-loader": "6.7.1",
-    "cssnano": "5.1.12",
-    "gh-pages": "6.3.0",
-    "html-webpack-plugin": "5.5.0",
-    "mini-css-extract-plugin": "2.6.1",
-    "postcss-loader": "7.0.1",
-    "style-loader": "3.3.1",
-    "webpack": "5.74.0",
-    "webpack-cli": "4.10.0",
-    "webpack-dev-server": "4.9.3"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/NoaTheGenii/se_project_spots.git"
-  },
-  "keywords": [
-    "\"photo-sharing\"",
-    "\"pictures\""
-  ],
-  "author": "Noa Dusett",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/NoaTheGenii/se_project_spots/issues"
-  },
-  "homepage": "https://github.com/NoaTheGenii/se_project_spots#readme",
-  "dependencies": {
-    "core-js": "3.43.0"
-  }
-}
