--- postcss.config.js
@@ -1,6 +1,0 @@
-const autoprefixer = require("autoprefixer");
-const cssnano = require("cssnano");
-
-module.exports = {
-    plugins: [autoprefixer, cssnano({ preset: "default" })],
-};
