--- src/blocks/profile.css
@@ -1,195 +1,0 @@
-.profile {
-  margin: 34px auto 20px;
-  padding-bottom: 20px;
-  max-width: 1280px;
-  line-height: 1;
-  box-sizing: border-box;
-  border-bottom: solid rgba(33, 33, 33, 0.4) 1px;
-  display: flex;
-  justify-content: space-between;
-}
-
-.profile__heading {
-  display: flex;
-  font-display: row;
-  gap: 20px;
-}
-
-.profile__avatar-container {
-  position: relative;
-  margin: 0;
-  padding: 0;
-  border-radius: 8px;
-  align-self: center;
-}
-
-.profile__avatar-btn {
-  position: absolute;
-  margin: 0;
-  padding: 0 0 12px;
-  width: 100%;
-  height: 100%;
-  font-size: 16px;
-  line-height: 1;
-  background-color: transparent;
-  border: none;
-  display: flex;
-  justify-content: center;
-  align-items: flex-end;
-  visibility: hidden;
-}
-
-.profile__avatar-container:hover .profile__avatar-btn {
-  visibility: visible;
-  cursor: pointer;
-}
-
-.profile__avatar-btn::after {
-  content: "Edit Profile";
-  color: #fff;
-}
-
-.profile__avatar {
-  width: 120px;
-  height: 120px;
-  border-radius: 8px;
-}
-
-.profile__column {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.profile__name {
-  margin: 0 0 12px;
-  padding: 0;
-  color: #212121;
-  font-size: 32px;
-  line-height: 1;
-  font-weight: 500;
-  text-overflow: ellipsis;
-  overflow: hidden;
-  line-clamp: 3;
-  display: -webkit-box;
-  -webkit-line-clamp: 3;
-  -webkit-box-orient: vertical;
-  max-width: 420px;
-}
-
-.profile__text {
-  margin: 0;
-  padding: 0;
-  font-size: 16px;
-  line-height: 1;
-  font-weight: 500;
-  color: rgba(33, 33, 33, 0.7);
-  text-overflow: ellipsis;
-  overflow: hidden;
-  line-clamp: 3;
-  display: -webkit-box;
-  -webkit-line-clamp: 3;
-  -webkit-box-orient: vertical;
-  max-width: 420px;
-}
-
-.profile__btn {
-  margin: 0;
-  padding: 0;
-  display: flex;
-  background-color: transparent;
-  border: none;
-}
-
-.profile__edit-btn {
-  margin: 0;
-  padding: 0;
-  height: 16px;
-  align-self: flex-start;
-  opacity: 0.7;
-  transition: 0.5s ease-out;
-}
-
-.profile__edit-btn:hover {
-  opacity: 1;
-  cursor: pointer;
-}
-
-.profile__add-img {
-  margin: 0;
-  padding: 0;
-  align-self: flex-end;
-  height: 52px;
-  transition: 0.5s ease-out;
-}
-
-.profile__add-img:hover {
-  opacity: 0.8;
-  cursor: pointer;
-}
-
-@media screen and (max-width: 627px) {
-  .profile {
-    margin: 20px 16px;
-    min-width: 288px;
-    flex-direction: column;
-    align-items: center;
-  }
-
-  .profile__heading {
-    margin-bottom: 20px;
-    flex-direction: column;
-    gap: 12px;
-  }
-
-  .profile__avatar {
-    height: 80px;
-    width: 80px;
-    display: flex;
-    align-self: center;
-  }
-
-  .profile__avatar-btn {
-    padding: 0;
-    left: 52px;
-    top: 52px;
-    width: 23.34px;
-    height: 23.34px;
-    display: block;
-    background-color: rgba(0, 0, 0, 0.6);
-    border-radius: 6px;
-    visibility: visible;
-  }
-
-  .profile__avatar-btn::after {
-    content: "";
-  }
-
-  .profile__name {
-    margin: 0;
-    max-width: 288px;
-    font-size: 20px;
-    line-height: 1;
-  }
-
-  .profile__pair {
-    margin-bottom: 32px;
-    text-align: center;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 6px;
-  }
-
-  .profile__btn {
-    display: flex;
-    align-self: center;
-  }
-
-  .profile__add-img {
-    background-image: url(../images/post-button-large.svg);
-    background-repeat: no-repeat;
-    align-self: center;
-    width: 288px;
-  }
-}
