--- src/utils/helpers.js
@@ -1,12 +1,0 @@
-export default function setButtonText(
-  button,
-  isLoading,
-  loadingText = "Saving...",
-  originalText = "Save"
-) {
-  if (isLoading) {
-    button.textContent = loadingText;
-  } else {
-    button.textContent = originalText;
-  }
-}
